# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ExceptionPage": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "description": "Records in this page.",
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExceptionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` for the next page. Null at end.",
            "title": "Next Cursor"
          },
          "total": {
            "description": "Total matching records.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "count",
          "items"
        ],
        "title": "ExceptionPage",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "An autonomous finance controller. Reconciles transactions across gateway, settlement, bank and internal ledger, and reports every record it refuses to guess at as a typed exception with the money at risk.\n\n**Money is always integer minor units** (paise for INR) paired with a currency \u2014 never a float, never pre-formatted. Format at the display edge.\n\nEvery endpoint requires a bearer token. Collections are cursor-paginated.",
    "title": "Ledger",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/audit": {
      "get": {
        "description": "Every automated decision, reproducible.\n\nRestricted to reviewers and above: the trail names counterparty references\nand amounts, which is more than a read-only viewer needs.\n\nThe `decision` filter matters more than it looks. A batch writes one entry per\nmatch, so refusals \u2014 the decisions that left money unreconciled and are the\nones an auditor actually came for \u2014 sit behind thousands of successes. Without\na filter they are technically present and practically unreachable.",
        "operationId": "list_audit_v1_audit_get",
        "parameters": [
          {
            "description": "Filter by decision, e.g. `match_refused` to see only where the engine declined to resolve something.",
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by decision, e.g. `match_refused` to see only where the engine declined to resolve something.",
              "title": "Decision"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Audit V1 Audit Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The decision trail for the current batch",
        "tags": [
          "audit"
        ]
      }
    },
    "/v1/batches": {
      "post": {
        "description": "Ingest, reconcile, raise exceptions, and score against ground truth.\n\nGated on `run_batch` rather than a write permission generally: this is the\nmost expensive operation the service offers, and rate limiting alone would\nnot stop an authenticated viewer from running it repeatedly.",
        "operationId": "run_batch_v1_batches_post",
        "parameters": [
          {
            "description": "Data seed. 1337 is the held-out set.",
            "in": "query",
            "name": "seed",
            "required": false,
            "schema": {
              "default": 1337,
              "description": "Data seed. 1337 is the held-out set.",
              "title": "Seed",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "events",
            "required": false,
            "schema": {
              "default": 2000,
              "maximum": 50000,
              "minimum": 1,
              "title": "Events",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run a reconciliation batch",
        "tags": [
          "batches"
        ]
      }
    },
    "/v1/batches/current": {
      "get": {
        "operationId": "current_batch_v1_batches_current_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summary of the current batch",
        "tags": [
          "batches"
        ]
      }
    },
    "/v1/exceptions": {
      "get": {
        "description": "The honest list. Sorted by money at risk, descending.\n\nTriage order is the point of this endpoint: a reviewer with an hour should\nspend it on the largest exposure, not on whatever the engine happened to\nraise first.",
        "operationId": "list_exceptions_v1_exceptions_get",
        "parameters": [
          {
            "description": "Filter by exception type.",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exception type.",
              "title": "Type"
            }
          },
          {
            "description": "Filter by severity.",
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity.",
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List unresolved exceptions",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/v1/exceptions/{exception_id}": {
      "get": {
        "operationId": "get_exception_v1_exceptions__exception_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "exception_id",
            "required": true,
            "schema": {
              "title": "Exception Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One exception, with its candidates",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/v1/exceptions/{exception_id}/resolve": {
      "post": {
        "description": "Record how an exception was settled.\n\nTwo permissions, deliberately. `resolve_exception` admits you here; supplying\n`accept_candidate_index` additionally requires `approve_money_movement`,\nbecause confirming a grouping asserts where money went. A reviewer can work\nthe queue; only a controller can say the money landed a particular way.\n\nIdempotent on `idempotency_key`: a retry is acknowledged, not re-applied.",
        "operationId": "resolve_exception_v1_exceptions__exception_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "exception_id",
            "required": true,
            "schema": {
              "title": "Exception Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve an exception",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/v1/matches": {
      "get": {
        "operationId": "list_matches_v1_matches_get",
        "parameters": [
          {
            "description": "Filter by cascade tier.",
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by cascade tier.",
              "title": "Tier"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List reconciled matches with their reasons",
        "tags": [
          "matches"
        ]
      }
    },
...
  }
}